simplify
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@6003 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ regression/cbmc/constant_folding2/main.c
@@ -7,12 +7,12 @@ typedef struct _pair
 int main (void)
 {
   pair p;
-  p.x = 0.0;
-  p.y = (int)0.0f;
+  p.x = 0;
+  p.y = 0;
   
   int array[2];
   array[0] = 0;
-  array[1] = (int)0.0f;
+  array[1] = 0;
   
   int i=0;
